Add dedicated header file for buttons command picker
